1.6.beta1 Is made for 64 bit and 32 bit Intel as well as PPC. Needs to be tested. Has "Flash x8 mode" to automatically multiply values by 8 when working with Flash.

+++ iHaxGamez/MyTableView.h
@@ -28,6 +28,6 @@
 
 - (void)awakeFromNib;
 
-- (void)editColumn:(int)columnIndex row:(int)rowIndex withEvent:(NSEvent *)theEvent select:(BOOL)flag;
+- (void)editColumn:(NSInteger)columnIndex row:(NSInteger)rowIndex withEvent:(NSEvent *)theEvent select:(BOOL)flag;
 - (void)textDidEndEditing:(NSNotification *)aNotification;
 @end
